% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSV.R
\name{AIC.stochvolTMB}
\alias{AIC.stochvolTMB}
\title{Calculate BIC of model}
\usage{
\method{AIC}{stochvolTMB}(object)
}
\arguments{
\item{object}{A \code{stochvolTMB} object.}
}
\value{
\code{BIC}: BIC of fitted model.
}
\description{
Calculate BIC of model
}
