% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSV.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Optimize negative log likelihood}
\usage{
estimate_parameters(data, method, opt.control = NULL,
  report = c("fixed", "report", "random"))
}
\arguments{
\item{opt.control}{An optional list of parameters for nlminb}

\item{report}{Vector of character}

\item{obj}{TMB object returned from \code{get_nll}.}
}
\description{
Optimize negative log likelihood
}
