% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSV.R
\name{get_nll}
\alias{get_nll}
\title{Construct objective function with derivatives using \link[TMB]{MakeADFun}}
\usage{
get_nll(data, model = "gaussian", ...)
}
\arguments{
\item{data}{vector of observations.}

\item{model}{string specifying distribution of error term in observational equation.}

\item{...}{additional arguments passed to \link[TMB]{MakeADFun}.}
}
\value{
List of components returned from \link[TMB]{MakeADFun}.
}
\description{
Construct objective function with derivatives using \link[TMB]{MakeADFun}
}
\keyword{internal}
