% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSV.R
\name{logit}
\alias{logit}
\title{Logit transformation from the real line to (-1, 1).}
\usage{
logit(x)
}
\arguments{
\item{x}{double}
}
\value{
double
}
\description{
Logit transformation from the real line to (-1, 1).
}
