% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSV.R
\name{optSV}
\alias{optSV}
\title{Function for parameter estimation of stochastic volatility models}
\usage{
optSV(data, method = "gaussian", report = c("fixed", "random"))
}
\arguments{
\item{data}{vector of observations}

\item{method}{string specifying distribution of error term in observational equation}
}
\value{
tibble of parameter estimates and standard error of estimates
}
\description{
Function for parameter estimation of stochastic volatility models
}
