% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volplot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Add predicted volatility.}
\usage{
plot_forecast(p, forecast, include_ci = TRUE)
}
\arguments{
\item{p}{ggplot object}

\item{forecast}{data.table}

\item{include_ci}{bool. Should 95 percent confidence interval be plotted?}
}
\value{
ggplot object
}
\description{
Adds predicted volatility to the volatility plot.
}
\keyword{internal}
