% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{Calculate one-step-ahead (OSA) residuals for stochastic volatility model.}
\usage{
residuals(object, conditional = 1:(object$nobs - 100), ...)
}
\arguments{
\item{object}{A \code{stochvolTMB} object.}

\item{...}{Currently not used.}
}
\value{
Vector of one-step-ahead residuals. If the model is correctly specified, these should be standard normal.
}
\description{
Calculate one-step-ahead (OSA) residuals for stochastic volatility model.
}
