% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSV.R
\name{simSV}
\alias{simSV}
\title{Simulate data for the stochastic volatility model}
\usage{
simSV(param, T = 1000, seed = NULL, method = "gaussian")
}
\arguments{
\item{param}{List of parameters}

\item{T}{Length of time series}

\item{seed}{Seed to reproduce simulation}

\item{method}{Distribtion of error term (only Gaussian so far)}
}
\description{
Simulate data for the stochastic volatility model
}
