% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spy}
\alias{spy}
\title{Daily closing prices for the S&P500 from 2005 to 2018.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{date}{date}
  \item{price}{price, in US dollars}
  \item{log_return}{logarithmic return}
  ...
}
}
\usage{
spy
}
\description{
A dataset containing the prices and log-returns of the S&P500 from 2005 to 2018
}
\keyword{datasets}
