\name{stochvolTMB-package}
\alias{stochvolTMB-package}
\alias{stochvolTMB}
\docType{package}
\title{
\packageTitle{stochvolTMB}
}
\description{
\packageDescription{stochvolTMB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stochvolTMB}
\packageIndices{stochvolTMB}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stochvolTMB}

Maintainer: \packageMaintainer{stochvolTMB}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
