% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSV.R
\name{summary.stochvolTMB}
\alias{summary.stochvolTMB}
\title{Summary tables of model parameters}
\usage{
\method{summary}{stochvolTMB}(object, ..., report = c("all", "fixed", "transformed", "random"))
}
\arguments{
\item{object}{A \code{stochvolTMB} object.}

\item{...}{Currently not used.}

\item{report}{Parameters to report with uncertainty estimates. Can be any subset of "fixed", "transformed" or 
"random" (see \link[TMB]{summary.sdreport}). "fixed" 
report the parameters on the scale they were estimated, for example are all standard deviations estimated on 
log scale. "transformed" report all transformed parameters, for example estimated standard deviations 
transformed from log scale by taking the exponential. Lastly, "random"
report the estimated latent log-volatility.}
}
\value{
\code{data.table} with parameter estimates, standard error, z-value and approximated p-value.
}
\description{
Extract parameters, transformed parameters and latent log volatility along with standard error, z-value and p-value
}
