% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSV.R
\name{summary.stochvolTMB}
\alias{summary.stochvolTMB}
\title{Summary tables of model parameters}
\usage{
\method{summary}{stochvolTMB}(object)
}
\description{
Extract parameters, transformed parameters and latent log volatility
}
