% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volplot.R
\name{volplot}
\alias{volplot}
\title{Function for plotting log volatiliy}
\usage{
volplot(report, include_ci = TRUE, plot_log = "TRUE")
}
\arguments{
\item{report}{tibble returned from optSV}

\item{include_ci}{logical. Include +- two times standard error}

\item{plot_log}{logical. Plot estimated log volatility. If false, exp(h / 2) is plotted.}
}
\value{
ggplot object
}
\description{
Function for plotting log volatiliy
}
