% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSV.R
\name{get_nll}
\alias{get_nll}
\title{Function for parameter estimation of stochastic volatility models}
\usage{
get_nll(data, model = "gaussian")
}
\arguments{
\item{data}{vector of observations}

\item{model}{string specifying distribution of error term in observational equation}
}
\value{
list of summary report and opt object
}
\description{
Function for parameter estimation of stochastic volatility models
}
