% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{Calculate one-step-ahead (OSA) residuals for stochastic volatility model.}
\usage{
residuals(obj)
}
\arguments{
\item{obj}{Output from optSV}
}
\value{
vector of one-step-ahead residuals of length \code{T}, where \code{T} is the number of observations.
}
\description{
Calculate one-step-ahead (OSA) residuals for stochastic volatility model.
}
