% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSV.R
\name{simSV}
\alias{simSV}
\title{Simulate data for the stochastic volatility model}
\usage{
simSV(param, N = 1000, seed = NULL, model = "gaussian")
}
\arguments{
\item{param}{List of parameters}

\item{N}{Length of time series}

\item{seed}{Seed to reproduce simulation}

\item{model}{Distribution of error term}
}
\value{
tibble of dimension \code{N x 2}
}
\description{
Simulate data for the stochastic volatility model
}
